[CMake] Install library
Install the library and the headers it depends on.
Works with a shared library build (`BUILD_SHARED_LIBS=ON`).

In theory it works with a static library build as well,
but the installation step would have to include the SPIR-V dependencies as static libraries,
which this patch doesn't address.

Provided some installation info.